[MRESOLVER-461] Cleanup in dep managers
3 of them were almost same.

---

https://issues.apache.org/jira/browse/MRESOLVER-461
